requirements.txt
crewai>=0.100.0
crewai-tools>=0.33.0
google-genai
python-dotenv
streamlit
